# Improve performance of VarZeroVec::deserialize and add provider benches