--- a/provider/testdata/Cargo.toml
+++ b/provider/testdata/Cargo.toml
@@ -325,6 +325,7 @@ icu_timezone = { version = "1.0.0-beta1", path = "../../components/timezone", de
 
 [dev-dependencies]
 icu = { path = "../../components/icu" }
+criterion = "0.3"
 
 [features]
 default = [
@@ -380,3 +381,7 @@ required-features = ["bin"]
 name = "icu4x-testdata-datagen"
 path = "src/bin/datagen.rs"
 required-features = ["bin"]
+
+[[bench]]
+name = "providers"
+harness = false