Fixes #2600

After @zbraniecki noticed a performance issue with StaticDataProvider's constructor, I looked into the code with @Manishearth  and found some low-hanging fruit to optimize. I managed to get a 63% improvement, as shown below with the new benches run across the second commit in this PR:

```
Benchmarking provider/construct/static: Collecting 100 samples in estimated 5.0030 s (5.2M iterati                                                                                                  provider/construct/static                        
                        time:   [951.26 ns 952.42 ns 953.85 ns]
                        change: [-63.724% -63.601% -63.474%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe

Benchmarking provider/construct/blob: Collecting 100 samples in estimated 5.4580 s (25k iterations                                                                                                  provider/construct/blob time:   [214.57 us 214.85 us 215.23 us]
                        change: [-0.7276% -0.5377% -0.3567%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe

Benchmarking provider/construct/baked: Collecting 100 samples in estimated 5.0000 s (18446744074B                                                                                                   provider/construct/baked                        
                        time:   [0.0000 ps 0.0000 ps 0.0000 ps]
                        change: [-45.396% -1.3066% +75.734%] (p = 0.98 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

provider/load/static    time:   [117.46 ns 117.92 ns 118.47 ns]                                 
                        change: [-0.8186% +1.4342% +4.4178%] (p = 0.35 > 0.05)
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  2 (2.00%) high mild
  15 (15.00%) high severe

provider/load/blob      time:   [114.53 ns 114.67 ns 114.83 ns]                               
                        change: [-3.5102% -3.2867% -3.0490%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high severe

provider/load/baked     time:   [39.669 ns 39.951 ns 40.354 ns]                                 
                        change: [+7.0914% +7.7782% +8.7248%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe
```