@sffc 2022-09-21
I included an extra commit, 508372c58e45c50e9044e7a8c6b964d9b8870366, with the following characteristics:

```
Benchmarking provider/construct/static: Collecting 100 samples in estimated 5.0000 s (7.0M iterati                                                                                                  provider/construct/static                        
                        time:   [707.11 ns 708.16 ns 709.41 ns]
                        change: [-25.849% -25.592% -25.343%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe

Benchmarking provider/construct/blob: Collecting 100 samples in estimated 5.4600 s (25k iterations                                                                                                  provider/construct/blob time:   [213.82 us 214.17 us 214.68 us]
                        change: [-0.5128% -0.3020% -0.0509%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  1 (1.00%) high severe

Benchmarking provider/construct/baked: Collecting 100 samples in estimated 5.0000 s (18446744074B                                                                                                   provider/construct/baked                        
                        time:   [0.0000 ps 0.0000 ps 0.0000 ps]
                        change: [-45.589% +0.0750% +89.388%] (p = 1.00 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

provider/load/static    time:   [133.00 ns 133.25 ns 133.56 ns]                                 
                        change: [+6.5930% +9.8142% +12.347%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe

provider/load/blob      time:   [132.53 ns 132.71 ns 132.92 ns]                               
                        change: [+15.493% +15.828% +16.180%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe

provider/load/baked     time:   [36.813 ns 37.003 ns 37.284 ns]                                 
                        change: [-8.6583% -7.8680% -7.1866%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
```

An extra 25% off the constructor at the cost of around 10% on the load function. It also saves a small amount of space in the postcard file. I am neutral on whether to include this additional change.