@robertbastian 2022-09-21
The additional change increases maintenance complexity for the blob schema. I'd rather keep complicated ULE stuff contained in the zerovec crate.

What's the core of that change, using a `FlexZeroVec` for the indices instead of a `&[u8]`? Why can this not be done in VZV?