@sffc 2022-09-21
> The additional change increases maintenance complexity for the blob schema. I'd rather keep complicated ULE stuff contained in the zerovec crate.
> 
> What's the core of that change, using a `FlexZeroVec` for the indices instead of a `&[u8]`? Why can this not be done in VZV?

See my comment threads in the code. The core of 508372c58e45c50e9044e7a8c6b964d9b8870366 is delaying the validation of the indices array; VZV requires that it be sorted, but it doesn't really need to be since we're not validating the byte slices; we just hand them off to Postcard to deal with.