@sffc 2022-09-21
I dropped 508372c58e45c50e9044e7a8c6b964d9b8870366 to get the uncontroversial change checked in, and I'll open another PR with the schema change to discuss further.